+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 6.1.0-alpha.2
 
 * Bump Sentry Android SDK to [5.2.0](https://github.com/getsentry/sentry-dart/pull/594) (#594)
   - [changelog](https://github.com/getsentry/sentry-java/blob/5.2.0/CHANGELOG.md)